third_party/WebKit/LayoutTests/fast/canvas/draw-custom-focus-ring.html

<!DOCTYPE HTML>
<head>
<title>Canvas test: drawCustomFocusRing</title>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body style="padding: 0; margin: 0">
<canvas id="canvas" class="output" width="300" height="350">
    <button id="button1"></button>
    <button id="button2"></button>
</canvas>
<script>
if (window.testRunner)
    testRunner.dumpAsText();

window.button1 = document.getElementById("button1");
window.button2 = document.getElementById("button2");

var canvas = document.getElementById("canvas").getContext("2d");
canvas.beginPath();
canvas.rect(50, 50, 200, 100);
canvas.fill();
shouldBe("canvas.drawCustomFocusRing(button1)", "false");
shouldBe("canvas.drawCustomFocusRing(button2)", "false");
shouldBe("button1.focus(); canvas.drawCustomFocusRing(button1)", "true");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
