third_party/WebKit/LayoutTests/fast/canvas/draw-custom-focus-ring-expected.txt

PASS canvas.drawCustomFocusRing(button1) is false
PASS canvas.drawCustomFocusRing(button2) is false
PASS button1.focus(); canvas.drawCustomFocusRing(button1) is true
PASS successfullyParsed is true

TEST COMPLETE
